tidy up what lives where in ir and ir-types
