+++ src/ir-types.h
@@ -300,7 +300,14 @@ typedef struct IrCtx {
     Map *labels;
 } IrCtx;
 
+/* Memory management */
+void irMemoryInit(void);
+void irMemoryRelease(void);
+void *irAlloc(u32 size);
+void irMemoryStats(void);
+
 IrCtx *irCtxNew(Cctrl *cc);
+void irCtxAddFunction(IrCtx *ctx, IrFunction *func);
 IrValue *irFnGetVar(IrFunction *func, u32 lvar_id);
 
 IrBlockMapping *irFunctionGetBlockMapping(IrFunction *func, IrBlock *ir_block);
@@ -328,5 +335,44 @@ IrBlock *irInstrGetTargetBlock(IrInstr *instr);
 IrBlock *irInstrGetFallthroughBlock(IrInstr *instr);
 Map *irBlockGetSuccessors(IrFunction *func, IrBlock *block);
 int irBlockIsStartOrEnd(IrFunction *func, IrBlock *block);
+
+/* Constructors */
+Vec *irFunctionVecNew(void);
+
+Vec *irPairVecNew(void);
+Vec *irValueVecNew(void);
+Map *irBlockMapNew(void);
+Map *irBlockMappingMapNew(void);
+IrBlockMapping *irBlockMappingNew(int id);
+Map *irVarValueMapNew(void);
+IrBlock *irBlockNew(void);
+IrValue *irTmp(IrValueType type, u16 size);
+IrInstr *irInstrNew(IrOp op, IrValue *dst, IrValue *r1, IrValue *r2);
+IrFunction *irFunctionNew(AoStr *fname);
+IrValue *irValueNew(IrValueType type, IrValueKind kind);
+
+void irFunctionAddMapping(IrFunction *func, IrBlock *src, IrBlock *dest);
+void irFunctionAddSuccessor(IrFunction *func, IrBlock *src, IrBlock *dest);
+void irFunctionAddPredecessor(IrFunction *func, IrBlock *src, IrBlock *prev);
+void irFunctionRemoveSuccessor(IrFunction *func, IrBlock *src, IrBlock *dest);
+void irFunctionRemovePredecessor(IrFunction *func, IrBlock *src, IrBlock *prev);
+
+void irResetBlockId(void);
+void irTmpVariableCountReset(void);
+
+void irBlockMappingRelease(void *_mapping);
+void irBlockRelease(IrBlock *block);
+void irFunctionRelease(IrFunction *func);
+
+
+
+
+
+void irFnAddVar(IrFunction *func, u32 lvar_id, IrValue *var);
+void irFnAddBlock(IrFunction *fn, IrBlock *block);
+int irSetVariable(IrFunction *func, u32 var_id, IrValue *var);
+void irAddStackSpace(IrCtx *ctx, int size);
+
+
 
 #endif
